FIX: prevent from unwanted enabling of the adjusting regime
The CTRL key status, which is hard to monitor when the slider is
disabled and which is now checked on every mouse event occasion,
should be updated only "one direction", only to monitor if CTRL
was depressed -- basically to monitor only if the CTRL-is-pressed
state still lasts. The reason is that we want CTRL to be pressed
explicitly/separately before the mouse event (to disable the slider
"in time" before the mouse event -- when mouse is pressed later,
the slider will already be disabled and thus immune to dragging of
its thumb(s)). So, noticing pressed CTRL already during mouse event,
should not start the slider adjusting regime (notice also the keyPressed()
refuses to start the regime if mouse button is already pressed) -- and
that is why the ctrl update goes only one direction.

# Conflicts:
#	src/main/java/bdv/util/ui/sliders/AbstractAdjustableSliderBasedControl.java